+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/datagen/datapack/worldgen/ModNoiseSettings.kt
@@ -157,5 +157,4 @@ object ModNoiseSettings {
 		return ResourceKey.create(Registries.NOISE_SETTINGS, ExcessiveUtilities.modResource(path))
 	}
 
-
 }